deck: splitted cards in multiple decks
splitted the uge cards module in a base
module with the prototypes and one module
per element.
